labs/python/complete/requirements.txt
#Flask==2.0.3
#ptvsd==4.3.2 # Required for debugging.
#google-cloud-spanner==3.13.0

Flask
gunicorn
google-cloud-spanner
